+++ core/src/main/resources/reference.conf
@@ -29,7 +29,7 @@ akka.persistence.cassandra {
   # see: https://docs.datastax.com/en/developer/java-driver/4.3/manual/core/configuration/#quick-overview
   # It may optionally have a constructor with an ActorSystem and Config parameter.
   # The config parameter is this config section of the plugin.
-  session-provider = "akka.cassandra.session.DefaultSessionProvider"
+  session-provider = "akka.stream.alpakka.cassandra.DefaultSessionProvider"
 
   session-name = ""
