+++ core/src/main/scala/akka/persistence/cassandra/journal/CassandraEventUpdate.scala
@@ -8,13 +8,14 @@ import akka.Done
 import akka.event.LoggingAdapter
 import akka.persistence.cassandra.PluginSettings
 import akka.persistence.cassandra.journal.CassandraJournal.{ Serialized, TagPidSequenceNr }
-import akka.cassandra.session.scaladsl.CassandraSession
 import com.datastax.oss.driver.api.core.cql.{ PreparedStatement, Row, Statement }
+
 import scala.collection.JavaConverters._
 import scala.concurrent.{ ExecutionContext, Future }
 import java.lang.{ Long => JLong }
 
 import akka.annotation.InternalApi
+import akka.stream.alpakka.cassandra.scaladsl.CassandraSession
 
 /** INTERNAL API */
 @InternalApi private[akka] trait CassandraEventUpdate {
