+++ core/src/main/scala/akka/persistence/cassandra/query/javadsl/CassandraReadJournal.scala
@@ -7,12 +7,12 @@ package akka.persistence.cassandra.query.javadsl
 import java.util.UUID
 import java.util.concurrent.CompletionStage
 
-import akka.cassandra.session.javadsl.CassandraSession
 import akka.{ Done, NotUsed }
 import akka.persistence.query.EventEnvelope
 import akka.persistence.query.Offset
 import akka.persistence.query.TimeBasedUUID
 import akka.persistence.query.javadsl._
+import akka.stream.alpakka.cassandra.javadsl.CassandraSession
 import akka.stream.javadsl.Source
 
 import scala.compat.java8.FutureConverters
