+++ core/src/main/scala/akka/persistence/cassandra/snapshot/CassandraSnapshotStatements.scala
@@ -10,9 +10,9 @@ import scala.concurrent.Future
 
 import akka.Done
 import akka.annotation.InternalApi
-import akka.cassandra.session.FutureDone
 import akka.persistence.cassandra.indent
 import com.datastax.oss.driver.api.core.CqlSession
+import akka.stream.alpakka.cassandra.FutureDone
 
 /**
  * INTERNAL API
