+++ core/src/test/scala/akka/persistence/cassandra/journal/CassandraEventUpdateSpec.scala
@@ -7,20 +7,19 @@ package akka.persistence.cassandra.journal
 import java.util.UUID
 
 import scala.concurrent.Await
-
 import akka.Done
 import akka.event.Logging
 import akka.persistence.PersistentRepr
 import akka.persistence.cassandra.journal.CassandraJournal.Serialized
-import akka.cassandra.session.scaladsl.CassandraSession
 import akka.persistence.cassandra.{ CassandraLifecycle, CassandraSpec, TestTaggingActor, _ }
 import akka.serialization.SerializationExtension
 import com.typesafe.config.ConfigFactory
+
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
-
 import akka.actor.ExtendedActorSystem
-import akka.cassandra.session.CqlSessionProvider
+import akka.stream.alpakka.cassandra.CqlSessionProvider
+import akka.stream.alpakka.cassandra.scaladsl.CassandraSession
 
 object CassandraEventUpdateSpec {
   val config = ConfigFactory.parseString("""
