+++ session/src/main/scala/akka/stream/alpakka/cassandra/package.scala
@@ -2,12 +2,12 @@
  * Copyright (C) 2016-2017 Lightbend Inc. <https://www.lightbend.com>
  */
 
-package akka.cassandra
+package akka.stream.alpakka
 
 import akka.Done
 
 import scala.concurrent.Future
 
-package object session {
+package object cassandra {
   val FutureDone: Future[Done] = Future.successful(Done)
 }
